notification_service/requirements/base.txt
# python-slugify==8.0.4  # https://github.com/un33k/python-slugify
# Pillow==11.0.0  # https://github.com/python-pillow/Pillow
# argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
# whitenoise==6.8.2  # https://github.com/evansd/whitenoise
# redis==5.2.0  # https://github.com/redis/redis-py
# hiredis==3.0.0  # https://github.com/redis/hiredis-py

# Django
# ------------------------------------------------------------------------------
django==5.0.9  # pyup: < 5.1  # https://www.djangoproject.com/
django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-model-utils==5.0.0  # https://github.com/jazzband/django-model-utils
# django-allauth[mfa]==65.1.0  # https://github.com/pennersr/django-allauth
# # Django REST Framework
djangorestframework==3.15.2  # https://github.com/encode/django-rest-framework
djangorestframework-simplejwt==5.3.1   # JWT token implementation for DRF
dj-rest-auth[with_social]==6.0.0       # REST API authentication utils
python-dotenv==1.0.1
django-extensions==3.2.1
django-filter==22.1
requests==2.26.0
django-celery-results==2.2.0
itsdangerous==2.2.0
pika
kombu
celery
django-celery-beat
importlib-metadata==6.5.0


